Clock implementation using threads in C language
Concurrent programming is achieved through multi-threading. Threads are the basic unit of execution in the processes. Threads generally have an id, program counter, Register set and Stack.
The above code implements a simple digital clock using pthread library of C which provides multi threading features.
Mutex and Semaphores are generally used for synchronization of the threads to solve the problem of the critical section in multi threading
A critical section is a section of code that is executed by multiple threads and where the sequence of execution for the threads makes a difference in the result of the concurrent execution of the critical section.